CASSSIDECAR-329:Fixed integration test failure for 4.0

+++ adapters/adapters-base/src/main/java/org/apache/cassandra/sidecar/adapters/base/jmx/GossipDependentStorageJmxOperations.java
@@ -215,7 +215,7 @@ public class GossipDependentStorageJmxOperations implements StorageJmxOperations
     }
 
     @Override
-    public long getCompactionThroughputMbPerSec()
+    public int getCompactionThroughputMbPerSec()
     {
         return delegate.getCompactionThroughputMbPerSec();
     }
