+++ server-common/src/main/java/org/apache/cassandra/sidecar/common/server/StorageOperations.java
@@ -173,5 +173,5 @@ public interface StorageOperations
      *
      * @return the current compaction throughput in megabytes per second, or 0 if throughput cannot be determined
      */
-    long getCompactionThroughputMbPerSec();
+    int getCompactionThroughputMbPerSec();
 }
